requirements.txt
python-telegram-bot
pyowm
requests
simplejson
